Change res.end to res.sendStatus
Calling `res.end` without any other actions on the `res` will send a response without a content type. Twilio will complain about that with a warning in the debugger.

`res.sendStatus(200);` will set the content type to `text/plain` and the body to "OK". It also has more intention in it ("I want to send a successful response").